chore(compiler): split up hir.rs into its own dir
I was working on #531 and #569, and adding even more impls to that giant file was super overwhelming. It's become unreadable! This is a bit of a bikeshed, but it splits up the HIR into separate files, similar to the `grammar` dir in the parser.
