+++ crates/apollo-compiler/src/database/hir/argument.rs
@@ -1,0 +1,45 @@
+use std::sync::Arc;
+
+use crate::hir::{HirNodeLocation, InputValueDefinition, Name, Value};
+
+#[derive(Clone, Debug, Hash, PartialEq, Eq)]
+pub struct Argument {
+    pub(crate) name: Name,
+    pub(crate) value: Value,
+    pub(crate) loc: HirNodeLocation,
+}
+
+impl Argument {
+    /// Get a reference to the argument's value.
+    pub fn value(&self) -> &Value {
+        &self.value
+    }
+
+    /// Get a reference to the argument's name.
+    pub fn name(&self) -> &str {
+        self.name.src()
+    }
+
+    /// Get the AST location information for this HIR node.
+    pub fn loc(&self) -> HirNodeLocation {
+        self.loc
+    }
+}
+
+#[derive(Clone, Debug, Hash, PartialEq, Eq)]
+pub struct ArgumentsDefinition {
+    pub(crate) input_values: Arc<Vec<InputValueDefinition>>,
+    pub(crate) loc: Option<HirNodeLocation>,
+}
+
+impl ArgumentsDefinition {
+    /// Get a reference to arguments definition's input values.
+    pub fn input_values(&self) -> &[InputValueDefinition] {
+        self.input_values.as_ref()
+    }
+
+    /// Get the AST location information for this HIR node.
+    pub fn loc(&self) -> Option<HirNodeLocation> {
+        self.loc
+    }
+}
