+++ crates/apollo-compiler/src/database/hir/directive.rs
@@ -1,0 +1,216 @@
+use std::{fmt, sync::Arc};
+
+use apollo_parser::ast;
+
+use crate::{
+    hir::{Argument, ArgumentsDefinition, HirNodeLocation, Name, Value},
+    HirDatabase,
+};
+
+#[derive(Clone, Debug, Hash, PartialEq, Eq)]
+pub struct Directive {
+    pub(crate) name: Name,
+    pub(crate) arguments: Arc<Vec<Argument>>,
+    pub(crate) loc: HirNodeLocation,
+}
+
+impl Directive {
+    /// Get a reference to the directive's name.
+    pub fn name(&self) -> &str {
+        self.name.src()
+    }
+
+    /// Get a reference to the directive's arguments.
+    pub fn arguments(&self) -> &[Argument] {
+        self.arguments.as_ref()
+    }
+
+    /// Get a reference to the value of the directive argument with the given name, if it exists.
+    pub fn argument_by_name(&self, name: &str) -> Option<&Value> {
+        Some(
+            self.arguments
+                .iter()
+                .find(|arg| arg.name() == name)?
+                .value(),
+        )
+    }
+
+    // Get directive definition of the currently used directive
+    pub fn directive(&self, db: &dyn HirDatabase) -> Option<Arc<DirectiveDefinition>> {
+        db.find_directive_definition_by_name(self.name().to_string())
+    }
+
+    /// Get the AST location information for this HIR node.
+    pub fn loc(&self) -> HirNodeLocation {
+        self.loc
+    }
+}
+
+#[derive(Clone, Debug, Hash, PartialEq, Eq)]
+pub struct DirectiveDefinition {
+    pub(crate) description: Option<String>,
+    pub(crate) name: Name,
+    pub(crate) arguments: ArgumentsDefinition,
+    pub(crate) repeatable: bool,
+    pub(crate) directive_locations: Arc<Vec<DirectiveLocation>>,
+    pub(crate) loc: HirNodeLocation,
+}
+
+impl DirectiveDefinition {
+    /// Get a reference to the directive definition's name.
+    pub fn name(&self) -> &str {
+        self.name.src()
+    }
+
+    /// Get a reference to Name's source.
+    pub fn name_src(&self) -> &Name {
+        &self.name
+    }
+
+    /// Get a reference to the directive definition's description.
+    pub fn description(&self) -> Option<&str> {
+        self.description.as_deref()
+    }
+
+    // Get a reference to argument definition's locations.
+    pub fn arguments(&self) -> &ArgumentsDefinition {
+        &self.arguments
+    }
+
+    // Get a reference to directive definition's locations.
+    pub fn directive_locations(&self) -> &[DirectiveLocation] {
+        self.directive_locations.as_ref()
+    }
+
+    /// Indicates whether a directive may be used multiple times in a single location.
+    pub fn repeatable(&self) -> bool {
+        self.repeatable
+    }
+
+    /// Get the AST location information for this HIR node.
+    pub fn loc(&self) -> HirNodeLocation {
+        self.loc
+    }
+
+    /// Get the location information for the "head" of the directive definition, namely the
+    /// `directive` keyword and the name.
+    pub(crate) fn head_loc(&self) -> HirNodeLocation {
+        self.name_src()
+            .loc()
+            .map(|name_loc| HirNodeLocation {
+                // Adjust the node length to include the name
+                node_len: name_loc.end_offset() - self.loc.offset(),
+                ..self.loc
+            })
+            .unwrap_or(self.loc)
+    }
+
+    /// Checks if current directive is one of built-in directives - `@skip`,
+    /// `@include`, `@deprecated`, `@specifiedBy`.
+    pub fn is_built_in(&self) -> bool {
+        matches!(
+            self.name(),
+            "skip" | "include" | "deprecated" | "specifiedBy"
+        )
+    }
+}
+
+#[derive(Copy, Clone, Debug, Hash, PartialEq, Eq)]
+pub enum DirectiveLocation {
+    Query,
+    Mutation,
+    Subscription,
+    Field,
+    FragmentDefinition,
+    FragmentSpread,
+    InlineFragment,
+    VariableDefinition,
+    Schema,
+    Scalar,
+    Object,
+    FieldDefinition,
+    ArgumentDefinition,
+    Interface,
+    Union,
+    Enum,
+    EnumValue,
+    InputObject,
+    InputFieldDefinition,
+}
+
+impl DirectiveLocation {
+    /// Get the name of this directive location as it would appear in GraphQL source code.
+    pub fn name(self) -> &'static str {
+        match self {
+            DirectiveLocation::Query => "QUERY",
+            DirectiveLocation::Mutation => "MUTATION",
+            DirectiveLocation::Subscription => "SUBSCRIPTION",
+            DirectiveLocation::Field => "FIELD",
+            DirectiveLocation::FragmentDefinition => "FRAGMENT_DEFINITION",
+            DirectiveLocation::FragmentSpread => "FRAGMENT_SPREAD",
+            DirectiveLocation::InlineFragment => "INLINE_FRAGMENT",
+            DirectiveLocation::VariableDefinition => "VARIABLE_DEFINITION",
+            DirectiveLocation::Schema => "SCHEMA",
+            DirectiveLocation::Scalar => "SCALAR",
+            DirectiveLocation::Object => "OBJECT",
+            DirectiveLocation::FieldDefinition => "FIELD_DEFINITION",
+            DirectiveLocation::ArgumentDefinition => "ARGUMENT_DEFINITION",
+            DirectiveLocation::Interface => "INTERFACE",
+            DirectiveLocation::Union => "UNION",
+            DirectiveLocation::Enum => "ENUM",
+            DirectiveLocation::EnumValue => "ENUM_VALUE",
+            DirectiveLocation::InputObject => "INPUT_OBJECT",
+            DirectiveLocation::InputFieldDefinition => "INPUT_FIELD_DEFINITION",
+        }
+    }
+}
+
+impl fmt::Display for DirectiveLocation {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        write!(f, "{}", self.name())
+    }
+}
+
+impl From<ast::DirectiveLocation> for DirectiveLocation {
+    fn from(directive_loc: ast::DirectiveLocation) -> Self {
+        if directive_loc.query_token().is_some() {
+            DirectiveLocation::Query
+        } else if directive_loc.mutation_token().is_some() {
+            DirectiveLocation::Mutation
+        } else if directive_loc.subscription_token().is_some() {
+            DirectiveLocation::Subscription
+        } else if directive_loc.field_token().is_some() {
+            DirectiveLocation::Field
+        } else if directive_loc.fragment_definition_token().is_some() {
+            DirectiveLocation::FragmentDefinition
+        } else if directive_loc.fragment_spread_token().is_some() {
+            DirectiveLocation::FragmentSpread
+        } else if directive_loc.inline_fragment_token().is_some() {
+            DirectiveLocation::InlineFragment
+        } else if directive_loc.variable_definition_token().is_some() {
+            DirectiveLocation::VariableDefinition
+        } else if directive_loc.schema_token().is_some() {
+            DirectiveLocation::Schema
+        } else if directive_loc.scalar_token().is_some() {
+            DirectiveLocation::Scalar
+        } else if directive_loc.object_token().is_some() {
+            DirectiveLocation::Object
+        } else if directive_loc.field_definition_token().is_some() {
+            DirectiveLocation::FieldDefinition
+        } else if directive_loc.argument_definition_token().is_some() {
+            DirectiveLocation::ArgumentDefinition
+        } else if directive_loc.interface_token().is_some() {
+            DirectiveLocation::Interface
+        } else if directive_loc.union_token().is_some() {
+            DirectiveLocation::Union
+        } else if directive_loc.enum_token().is_some() {
+            DirectiveLocation::Enum
+        } else if directive_loc.enum_value_token().is_some() {
+            DirectiveLocation::EnumValue
+        } else if directive_loc.input_object_token().is_some() {
+            DirectiveLocation::InputObject
+        } else {
+            DirectiveLocation::InputFieldDefinition
+        }
+    }
+}
