+++ crates/apollo-compiler/src/database/hir/variable.rs
@@ -1,0 +1,58 @@
+use std::sync::Arc;
+
+use crate::hir::{DefaultValue, Directive, HirNodeLocation, Name, Type};
+
+#[derive(Clone, Debug, Hash, PartialEq, Eq)]
+pub struct VariableDefinition {
+    pub(crate) name: Name,
+    pub(crate) ty: Type,
+    pub(crate) default_value: Option<DefaultValue>,
+    pub(crate) directives: Arc<Vec<Directive>>,
+    pub(crate) loc: HirNodeLocation,
+}
+
+impl VariableDefinition {
+    /// Get a reference to the variable definition's name.
+    pub fn name(&self) -> &str {
+        self.name.src()
+    }
+
+    /// Get a reference to the variable definition's ty.
+    pub fn ty(&self) -> &Type {
+        &self.ty
+    }
+
+    /// Get a reference to the variable definition's default value.
+    pub fn default_value(&self) -> Option<&DefaultValue> {
+        self.default_value.as_ref()
+    }
+
+    /// Get a reference to the variable definition's directives.
+    pub fn directives(&self) -> &[Directive] {
+        self.directives.as_ref()
+    }
+
+    /// Returns the first directive with the given name.
+    ///
+    /// For repeatable directives, see [`directives_by_name`][Self::directives_by_name] (plural).
+    pub fn directive_by_name(&self, name: &str) -> Option<&Directive> {
+        self.directives_by_name(name).next()
+    }
+
+    /// Returns an iterator of directives with the given name.
+    ///
+    /// For non-repeatable directives, [`directive_by_name`][Self::directive_by_name] (singular).
+    pub fn directives_by_name<'def: 'name, 'name>(
+        &'def self,
+        name: &'name str,
+    ) -> impl Iterator<Item = &'def Directive> + 'name {
+        self.directives()
+            .iter()
+            .filter(move |directive| directive.name() == name)
+    }
+
+    /// Get the AST location information for this HIR node.
+    pub fn loc(&self) -> HirNodeLocation {
+        self.loc
+    }
+}
